+++ openlineplanner/src/components/LineElement.vue
@@ -19,6 +19,13 @@
       />
       <span v-else class="grow">{{ line.name }}</span>
       <TooltipButton
+        v-if="editStore.isEditing == line"
+        toolTip="Remove line (permanent)"
+        :handler="removeLine"
+      >
+        <span class="line-element__remove"><TrashCanOutlineIcon /></span>
+      </TooltipButton>
+      <TooltipButton
         v-if="editStore.isEditing != line"
         :handler="toggleEditing"
         toolTip="Start editing line"
@@ -27,25 +34,6 @@
         <span>
           <PencilOutlineIcon />
         </span>
-      </TooltipButton>
-      <TooltipButton
-        v-if="
-          (editStore.isEditing == line &&
-            linesStore.lines[line.id].pointIds.length > 1) ||
-          findStationLoading
-        "
-        :handler="findStation"
-        toolTip="Automatically find ideal station"
-      >
-        <LoadingIcon v-if="findStationLoading" class="loader" />
-        <BusStopIcon v-else />
-      </TooltipButton>
-      <TooltipButton
-        v-if="editStore.isEditing == line"
-        toolTip="Remove line (permanent)"
-        :handler="removeLine"
-      >
-        <span class="line-element__remove"><TrashCanOutlineIcon /></span>
       </TooltipButton>
     </div>
     <ColorPicker
@@ -61,12 +49,13 @@
     />
     <div
       class="line-element__warning"
-      v-if="editStore.isEditing === line && editStore.isExtending != null"
+      v-if="editStore.isEditing === line"
+      @click="disableEditing"
     >
       <span class="line-element__warning__title"
-        >{{ line.name }} is being extended!</span
+        >{{ line.name }} is being {{ editStore.isExtending ? "extended": "edited"}}!</span
       >
-      <span>To end the line stop editing (Enter) <ArrowDownRightIcon /></span>
+      <span>Click here to stop {{ editStore.isExtending ? "extending  (Enter)": "editing"}}</span>
     </div>
   </div>
 </template>
@@ -74,9 +63,6 @@
 <script>
 import PencilOutlineIcon from "vue-material-design-icons/PencilOutline.vue";
 import TrashCanOutlineIcon from "vue-material-design-icons/TrashCanOutline.vue";
-import ArrowDownRightIcon from "vue-material-design-icons/ArrowDownRight.vue";
-import BusStopIcon from "vue-material-design-icons/BusStop.vue";
-import LoadingIcon from "vue-material-design-icons/Loading.vue";
 import TooltipButton from "./TooltipButton.vue";
 import IconLine from "./IconLine.vue";
 import { useLinesStore } from "../stores/lines";
@@ -92,15 +78,12 @@ export default {
   },
   components: {
     IconLine,
-    BusStopIcon,
     PencilOutlineIcon,
-    TrashCanOutlineIcon,
     ColorPicker,
-    ArrowDownRightIcon,
-    LoadingIcon,
     TooltipButton,
     TypePicker,
     TypeIcon,
+    TrashCanOutlineIcon,
   },
   data() {
     return {
@@ -109,10 +92,22 @@ export default {
       paxStore: usePaxStore(),
       selectColor: false,
       selectType: false,
-      findStationLoading: false,
     };
   },
+  mounted() {
+    window.addEventListener("keydown", (e) => {
+      if (this.editStore.isExtending) {
+        if (e.key === "Enter") {
+          this.disableEditing();
+        }
+      }
+    });
+  },
   methods: {
+    disableEditing() {
+      this.editStore.isEditing = null;
+      this.editStore.isExtending = null;
+    },
     editName(e) {
       this.linesStore.getLineById(this.line.id).name = e.srcElement.value;
     },
@@ -166,73 +161,6 @@ export default {
       this.editStore.isExtending = null;
       this.linesStore.removeLine(this.line);
     },
-    findStation() {
-      // Matomo Tracking
-      if (window && window.Piwik) {
-        window.Piwik.getTracker().trackEvent("editing", "findStation");
-      }
-      if (!this.findStationLoading) {
-        this.findStationLoading = true;
-        const route = [];
-        const stations = [];
-        const points = Object.values(this.linesStore.points);
-        if (points.length <= 0) {
-          return;
-        }
-        // ToDo: Add max coverage
-        points.forEach((point) => {
-          if (point.type === "station") {
-            stations.push({
-              lat: point.lat,
-              lng: point.lng,
-              id: point.id,
-              // Add max coverage
-              coverage: Math.max(
-                ...point.lines.map((lineId) =>
-                  this.linesStore.getLineById(lineId).getCoverage()
-                )
-              ),
-            });
-          }
-        });
-        this.linesStore
-          .getLineById(this.line.id)
-          .pointIds.forEach((pointRef) => {
-            const point = this.linesStore.getPointById(pointRef);
-            route.push({
-              lat: point.lat,
-              lng: point.lng,
-              id: point.id,
-            });
-          });
-
-        fetch(import.meta.env.VITE_API_ENDPOINT + "/find-station", {
-          method: "POST",
-          body: JSON.stringify({
-            route,
-            stations,
-            coverage: this.editStore.isEditing.getCoverage(),
-            method: this.paxStore.calculationMethod,
-          }),
-          headers: {
-            "Content-type": "application/json",
-          },
-        })
-          .then((data) => data.json())
-          .then((stationProposal) => {
-            const newPoint = this.linesStore.addPoint(
-              stationProposal.optimalStation.lat,
-              stationProposal.optimalStation.lng,
-              this.line,
-              stationProposal.optimalStation.index
-            );
-            newPoint.type = "station";
-          })
-          .finally(() => {
-            this.findStationLoading = false;
-          });
-      }
-    },
   },
 };
 </script>
@@ -241,7 +169,6 @@ export default {
 .line-element {
   display: flex;
   flex-direction: column;
-  margin-bottom: $space-ssm;
 
   &__data {
     display: flex;
@@ -260,7 +187,7 @@ export default {
     flex-direction: column;
     align-items: center;
     background-color: $c-accent-one;
-    // color: $c-text-pr;
+    cursor: pointer;
     padding: $space-ssm $space-ssm;
 
     &__title {
